requirements.txt
django==2.2.4
djangorestframework==3.10.2